elpaiscom-bot/requirements.txt
pika
beautifulsoup4
requests
pytz
feedparser
pymongo
python-dateutil